+++ kubejs/server_scripts/Mods/Greate/CrushingWheels.js
@@ -1,0 +1,3 @@
+ServerEvents.recipes((event) => {
+  event.remove({ type: "greate:crushing", id: /.*wood_chip_.*/ });
+});